[ADD] Null 처리 추가
