+++ smeem-api/src/main/java/com/smeem/api/member/api/dto/response/MemberGetResponse.java
@@ -1,12 +1,14 @@
 package com.smeem.api.member.api.dto.response;
 
-
 import com.smeem.api.badge.api.dto.response.BadgeResponse;
 import com.smeem.api.member.service.dto.response.MemberGetServiceResponse;
-import lombok.AccessLevel;
+import com.smeem.api.member.service.dto.response.MemberGetServiceResponse.PlanServiceResponse;
 import lombok.Builder;
 
-@Builder(access = AccessLevel.PRIVATE)
+import static java.util.Objects.nonNull;
+import static lombok.AccessLevel.PRIVATE;
+
+@Builder(access = PRIVATE)
 public record MemberGetResponse(
         String username,
         String target,
@@ -33,5 +35,21 @@ public record MemberGetResponse(
                 .badge(BadgeResponse.from(response.badge()))
                 .trainingPlan(PlanResponse.from(response.trainingPlan()))
                 .build();
+    }
+
+    @Builder(access = PRIVATE)
+    private record PlanResponse(
+            long id,
+            String content
+    ) {
+
+        private static PlanResponse from(PlanServiceResponse response) {
+            return nonNull(response)
+                    ? PlanResponse.builder()
+                        .id(response.id())
+                        .content(response.content())
+                        .build()
+                    : null;
+        }
     }
 }
